Update dependencies in bun.lock and package.json, add Zustand for state management, and introduce new stores for table management. Refactor URL sync functionality and enhance table store capabilities with new methods for state management.
