+++ packages/core/package.json
@@ -29,8 +29,9 @@
     "clean": "git clean -xdf .cache dist .turbo node_modules"
   },
   "dependencies": {
-    "date-fns": "^2.30.0",
-    "lz-string": "^1.5.0"
+    "date-fns": "4.1.0",
+    "lz-string": "^1.5.0",
+    "zustand": "^5.0.2"
   },
   "peerDependencies": {
     "react": "catalog:",
